📦 chore(ci.yml): add CI workflow for linting, building, and typechecking the project
This commit adds a new file `.github/workflows/ci.yml` which contains the CI workflow configuration. The workflow is triggered on every push event.

The workflow consists of a single job named "Lint" which runs on the latest version of Ubuntu. It sets the `CI` environment variable to `true`.

The steps in the job include:
- Checking out the repository
- Setting up pnpm package manager
- Setting up Node.js using the version specified in the `.tool-versions` file and caching the pnpm dependencies
- Installing project dependencies using `pnpm install --frozen-lockfile`
- Building the project using `pnpm build`
- Running typechecking using `pnpm typecheck`
- Running linting using `pnpm lint`